requirements.txt
discord.py==1.7.3
discord-py-slash-command==2.6.0